Converted firmware

+++ firmware/sources/hardware/storage/sdcard_storage.cpp
@@ -1,7 +1,7 @@
 // ***************************************************************************************
 // ***************************************************************************************
 //
-//      Name :      sdcard_storage.c
+//      Name :      sdcard_storage.cpp
 //      Author :    Carl John Kugler III
 //                  Paul Robson
 //      Date :      20th December 2023
@@ -60,8 +60,8 @@ static sd_card_t sd_cards[] = {  // One for each SD card
         .spi = &spis[0],  // Pointer to the SPI driving this card
         .ss_gpio = 25,    // The SPI slave select GPIO for this SD card
         .use_card_detect = false,
-        .card_detect_gpio = -1,   // Card detect
-        .card_detected_true = -1,  // What the GPIO read returns when a card is
+        .card_detect_gpio = (uint) -1,   // Card detect
+        .card_detected_true = (uint) -1,  // What the GPIO read returns when a card is
                                  // present. Use -1 if there is no card detect.
  
         .m_Status = STA_NOINIT
